Allow restart after refreshing the secret
See issue #2. This change can reduce the effective startup time of a
controller after its first installation from ~100s to ~10s, because we
no longer wait for the kubelet to project the new value of the secret.

This is almost identical to the code in HNC (see
https://github.com/kubernetes-sigs/multi-tenancy/pull/836). The only
difference is that this change restarts in _both_ codepaths where the
secrets are refreshed, while HNC only restarts when the secret
previously contained no valid data at all. This is almost certainly
unnecessary and has no impact, but is easier to understand.

Tested: the HNC code has worked reliably for about a month. While this
change is slightly different to that of HNC, note that the feature is
opt-in and initially I only expect HNC to use it. So I think this
sufficient testing to release a very early version of cert-controller,
import it into HNC, and test it there over the long term.

Signed-off-by: Adrian Ludwin <aludwin@google.com>